+++ CHANGELOG.md
@@ -51,6 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3223](https://github.com/open-telemetry/opentelemetry-python/pull/3223))
 - Add speced out environment variables and arguments for BatchLogRecordProcessor
   ([#3237](https://github.com/open-telemetry/opentelemetry-python/pull/3237))
+- Add benchmark tests for metrics
+  ([#3267](https://github.com/open-telemetry/opentelemetry-python/pull/3267))
 
 
 ## Version 1.17.0/0.38b0 (2023-03-22)